requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
defusedxml==0.7.1
idna==3.11
jira==3.10.5
oauthlib==3.3.1
packaging==25.0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
typing_extensions==4.15.0
urllib3==2.6.3
